Move arch_from_id itself into __init__.py, above __all__
The previous commit left a wrapper in __init__.py that tried the language id and
then called archinfo.arch.arch_from_id for everything else. Move the whole
function instead, so there is one definition and no delegation, and put it above
the __all__ list rather than after it.

get_host_arch moves with it: it is a one-line call to arch_from_id, and leaving
it behind would have needed a deferred import back into the package root, which
is the shape this is removing. arch.py keeps arch_id_map, register_arch and
ArchNotFound, and loses its now-unused platform import.

Nothing in the ecosystem imports either name from archinfo.arch; every caller
uses archinfo.arch_from_id.
